move CommitRepo to Scm namespace
its not Entity class
