+++ src/Scm/Adapter/Standard.php
@@ -16,6 +16,7 @@ namespace Eventum\Scm\Adapter;
 use Eventum\Model\Entity;
 use Eventum\Model\Repository\CommitRepository;
 use Eventum\Scm\Payload\StandardPayload;
+use Eventum\Scm\ScmRepository;
 use InvalidArgumentException;
 use Issue;
 use Symfony\Component\HttpFoundation\Request;
@@ -52,7 +53,7 @@ class Standard extends AbstractAdapter
         }
 
         $ci = $payload->createCommit();
-        $repo = new Entity\CommitRepo($ci->getScmName());
+        $repo = new ScmRepository($ci->getScmName());
 
         if (!$repo->branchAllowed($payload->getBranch())) {
             throw new InvalidArgumentException("Branch not allowed: {$payload->getBranch()}");
